+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/aktivitetsplikt/AktivitetspliktTidslinje.tsx
@@ -1,11 +1,7 @@
 import { HStack, VStack } from '@navikt/ds-react'
-import { hentAktiviteterOgHendelser } from '~shared/api/aktivitetsplikt'
 import { IDetaljertBehandling } from '~shared/types/IDetaljertBehandling'
-import { useApiCall } from '~shared/hooks/useApiCall'
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import { IAktivitetHendelse, IAktivitetPeriode } from '~shared/types/Aktivitetsplikt'
-import { isFailureHandler } from '~shared/api/IsFailureHandler'
-import { AktivitetOgHendelse } from '~components/behandling/aktivitetsplikt/AktivitetOgHendelse'
 import { AktivitetspliktTimeline } from '~components/behandling/aktivitetsplikt/aktivitetspliktTimeline/AktivitetspliktTimeline'
 
 export enum AktivitetspliktSkjemaAaVise {
@@ -32,27 +28,9 @@ interface Props {
 }
 
 export const AktivitetspliktTidslinje = ({ behandling, doedsdato, sakId }: Props) => {
-  const [hentAktivitetOgHendelserResult, hentAktiviteterOgHendelserRequest] = useApiCall(hentAktiviteterOgHendelser)
-  const [aktivitetPerioder, setAktivitetPerioder] = useState<IAktivitetPeriode[]>([])
-  const [aktivitetHendelser, setAktivitetHendelser] = useState<IAktivitetHendelse[]>([])
-  const [redigerAktivitet, setRedigerAktivitet] = useState<IAktivitetPeriode | undefined>(undefined)
-  const [redigerHendelse, setRedigerHendelse] = useState<IAktivitetHendelse | undefined>(undefined)
-
   const [aktivitetspliktRedigeringModus, setAktivitetspliktRedigeringModus] = useState<AktivitetspliktRedigeringModus>(
     defaultAktivitetspliktRedigeringModus
   )
-
-  useEffect(() => {
-    hentAktiviteterOgHendelserRequest({ sakId: sakId, behandlingId: behandling?.id }, (aktiviteter) => {
-      setAktivitetPerioder(aktiviteter.perioder)
-      setAktivitetHendelser(aktiviteter.hendelser)
-    })
-  }, [behandling, sakId])
-
-  function avbrytRedigering() {
-    setRedigerAktivitet(undefined)
-    setRedigerHendelse(undefined)
-  }
 
   return (
     <VStack gap="8" minWidth="50rem">
@@ -60,29 +38,21 @@ export const AktivitetspliktTidslinje = ({ behandling, doedsdato, sakId }: Props
         behandling={behandling}
         sakId={sakId}
         doedsdato={doedsdato}
-        aktivitetHendelser={aktivitetHendelser}
-        setAktivitetHendelser={setAktivitetHendelser}
-        aktivitetPerioder={aktivitetPerioder}
-        setAktivitetPerioder={setAktivitetPerioder}
         setAktivitetspliktRedigeringModus={setAktivitetspliktRedigeringModus}
       />
 
       <HStack align="center" justify="space-between">
-        <AktivitetOgHendelse
-          key={redigerAktivitet?.id}
-          behandling={behandling}
-          oppdaterAktiviteter={setAktivitetPerioder}
-          redigerAktivitet={redigerAktivitet}
-          sakId={sakId}
-          setHendelser={setAktivitetHendelser}
-          redigerHendelse={redigerHendelse}
-          avbrytRedigering={avbrytRedigering}
-        />
+        {/*<AktivitetOgHendelse*/}
+        {/*  key={redigerAktivitet?.id}*/}
+        {/*  behandling={behandling}*/}
+        {/*  oppdaterAktiviteter={setAktivitetPerioder}*/}
+        {/*  redigerAktivitet={redigerAktivitet}*/}
+        {/*  sakId={sakId}*/}
+        {/*  setHendelser={setAktivitetHendelser}*/}
+        {/*  redigerHendelse={redigerHendelse}*/}
+        {/*  avbrytRedigering={avbrytRedigering}*/}
+        {/*/>*/}
       </HStack>
-      {isFailureHandler({
-        errorMessage: 'En feil oppsto ved henting av tidslinje',
-        apiResult: hentAktivitetOgHendelserResult,
-      })}
     </VStack>
   )
 }
