+++ apps/etterlatte-saksbehandling-ui/client/src/components/behandling/aktivitetsplikt/aktivitetspliktTimeline/AktivitetHendelseTimelinePinContent.tsx
@@ -1,0 +1,93 @@
+import { IAktivitetHendelse } from '~shared/types/Aktivitetsplikt'
+import { BodyShort, Button, HStack, VStack } from '@navikt/ds-react'
+import { formaterDatoMedTidspunkt } from '~utils/formatering/dato'
+import { ReactNode } from 'react'
+import { useApiCall } from '~shared/hooks/useApiCall'
+import { slettAktivitetHendelseForBehandling, slettAktivitetHendelseForSak } from '~shared/api/aktivitetsplikt'
+import { PencilIcon, TrashIcon } from '@navikt/aksel-icons'
+import { isPending } from '~shared/api/apiUtils'
+import { isFailureHandler } from '~shared/api/IsFailureHandler'
+import {
+  AktivitetspliktRedigeringModus,
+  AktivitetspliktSkjemaAaVise,
+} from '~components/behandling/aktivitetsplikt/AktivitetspliktTidslinje'
+
+interface Props {
+  behandlingId?: string
+  sakId: number
+  aktivitetHendelse: IAktivitetHendelse
+  setAktivitetspliktRedigeringModus: (aktivitetspliktRedigeringModus: AktivitetspliktRedigeringModus) => void
+}
+
+export const AktivitetHendelseTimelinePinContent = ({
+  aktivitetHendelse,
+  behandlingId,
+  sakId,
+  setAktivitetspliktRedigeringModus,
+}: Props): ReactNode => {
+  const [slettAktivitetsHendelseForBehandlingResult, slettAktivitetsHendelseForBehandlingRequest] = useApiCall(
+    slettAktivitetHendelseForBehandling
+  )
+  const [slettAktivitetsHendelseForSakResult, slettAktivitetsHendelseForSakRequest] =
+    useApiCall(slettAktivitetHendelseForSak)
+
+  const slettAktivitetHendelse = (aktivitetHendelseId: string) => {
+    if (!!behandlingId) {
+      slettAktivitetsHendelseForBehandlingRequest({ behandlingId, aktivitetHendelseId })
+    } else {
+      slettAktivitetsHendelseForSakRequest({ sakId, aktivitetHendelseId })
+    }
+  }
+
+  return (
+    <VStack gap="2">
+      <BodyShort>{aktivitetHendelse.beskrivelse}</BodyShort>
+      <VStack>
+        <BodyShort>
+          <i>
+            Lagt til {formaterDatoMedTidspunkt(new Date(aktivitetHendelse.opprettet.tidspunkt))} av{' '}
+            {aktivitetHendelse.opprettet.ident}
+          </i>
+        </BodyShort>
+        <BodyShort>
+          <i>
+            Sist endret {formaterDatoMedTidspunkt(new Date(aktivitetHendelse.opprettet.tidspunkt))} av{' '}
+            {aktivitetHendelse.opprettet.ident}
+          </i>
+        </BodyShort>
+      </VStack>
+      {isFailureHandler({
+        apiResult: slettAktivitetsHendelseForBehandlingResult || slettAktivitetsHendelseForSakResult,
+        errorMessage: 'Kunne ikke slette aktivitet hendelse',
+      })}
+      <HStack gap="2">
+        <Button
+          variant="secondary"
+          size="xsmall"
+          icon={<PencilIcon aria-hidden />}
+          iconPosition="right"
+          loading={isPending(slettAktivitetsHendelseForBehandlingResult || slettAktivitetsHendelseForSakResult)}
+          onClick={() =>
+            setAktivitetspliktRedigeringModus({
+              aktivitetspliktSkjemaAaVise: AktivitetspliktSkjemaAaVise.AKTIVITET_HENDELSE,
+              aktivitetHendelse: aktivitetHendelse,
+              aktivitetPeriode: undefined,
+            })
+          }
+        >
+          Rediger
+        </Button>
+        <Button
+          variant="secondary"
+          size="xsmall"
+          icon={<TrashIcon aria-hidden />}
+          iconPosition="right"
+          loading={isPending(slettAktivitetsHendelseForBehandlingResult || slettAktivitetsHendelseForSakResult)}
+          onClick={() => slettAktivitetHendelse(aktivitetHendelse.id)}
+        >
+          Slett
+        </Button>
+      </HStack>
+    </VStack>
+  )
+}
